add create-app*
create-app* accepts an extra argument called routes that is an
association list which maps route patterns to controllers.

The route pattern can contain :foo path component, that will match the
what is in the url e.g. /foo/:bar/baz will match /foo/42/baz. The
model will be updated to contain both the pattern and the matched
parameters inside 'location respectively as 'path and 'location.

The associated controller allows to update the controller depending
on the route.